refactor: refactor the code
