+++ app/src/main/java/com/mudassir/giphyapi/di/modules/FragmentModule.kt
@@ -1,7 +1,7 @@
 package com.mudassir.giphyapi.di.modules
 
-import com.mudassir.giphyapi.ui.FavouritesFragment
-import com.mudassir.giphyapi.ui.GiphyTrendingFragment
+import com.mudassir.giphyapi.ui.fragments.FavouritesFragment
+import com.mudassir.giphyapi.ui.fragments.GiphyTrendingFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
