+++ app/src/main/java/com/mudassir/giphyapi/di/modules/ViewModelFactoryModule.kt
@@ -1,8 +1,7 @@
 package com.mudassir.giphyapi.di.modules
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.mudassir.giphyapi.ui.GiphyTrendingViewModel
+import com.mudassir.giphyapi.ui.viewModel.GiphyTrendingViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
@@ -10,8 +9,8 @@ import dagger.multibindings.IntoMap
 @Module
 abstract class ViewModelFactoryModule {
 
-    @Binds
-    abstract fun bindViewModelFactory(viewModelFactory: ViewModelFactory): ViewModelProvider.Factory
+//    @Binds
+//    abstract fun bindViewModelFactory(viewModelFactory: ViewModelFactory): ViewModelProvider.Factory
 
     @Binds
     @IntoMap
