+++ app/src/main/java/com/mudassir/giphyapi/ui/fragments/FavouritesFragment.kt
@@ -1,4 +1,4 @@
-package com.mudassir.giphyapi.ui
+package com.mudassir.giphyapi.ui.fragments
 
 import android.os.Bundle
 import android.util.Log
@@ -6,7 +6,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -15,6 +14,7 @@ import com.mudassir.giphyapi.databinding.FragmentFavouritesBinding
 import com.mudassir.giphyapi.di.modules.GenericSavedStateViewModelFactory
 import com.mudassir.giphyapi.di.modules.GiphyViewModelFactory
 import com.mudassir.giphyapi.ui.adapter.FavouriteAdapter
+import com.mudassir.giphyapi.ui.viewModel.GiphyTrendingViewModel
 import dagger.android.support.AndroidSupportInjection
 import javax.inject.Inject
