+++ app/src/main/java/com/mudassir/giphyapi/util/Constants.kt
@@ -1,4 +1,4 @@
-package com.mudassir.giphyapi
+package com.mudassir.giphyapi.util
 
 object Constants {
      const val SAVED_QUERY_KEY: String = "query"
