+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/trendingGiphyFragment"
-        android:name="com.mudassir.giphyapi.ui.GiphyTrendingFragment"
+        android:name="com.mudassir.giphyapi.ui.fragments.GiphyTrendingFragment"
         android:label="fragment_trending_giphy"
         tools:layout="@layout/fragment_trending_giphy" >
         <action
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/favouritesFragment"
-        android:name="com.mudassir.giphyapi.ui.FavouritesFragment"
+        android:name="com.mudassir.giphyapi.ui.fragments.FavouritesFragment"
         android:label="FavouritesFragment"
         tools:layout="@layout/fragment_favourites"
         />
