--- domain/src/main/java/com/mudassir/domain/mapper/GiphyItemMapper.kt
@@ -1,7 +1,0 @@
-package com.mudassir.domain.mapper
-
-import com.mudassir.core.util.Mapper
-import javax.inject.Inject
-
-//class GiphyItemMapper @Inject constructor(): Mapper<GiphyR> {
-//}
