+++ domain/src/main/java/com/mudassir/domain/usecase/GiphyTrendingUseCase.kt
@@ -7,25 +7,20 @@ import com.mudassir.core.util.ErrorFactory
 import com.mudassir.domain.model.GiphyDomainModel
 import com.mudassir.domain.repository.GiphyRepository
 import javax.inject.Inject
-import kotlin.math.log
 
-class GiphyTrendingUseCase @Inject constructor(private val giphyRepository: GiphyRepository,
-                                               private val errorFactory: ErrorFactory
-) : UseCase.UseCaseWithParam<String?,List<GiphyDomainModel>>  {
+class GiphyTrendingUseCase @Inject constructor(
+    private val giphyRepository: GiphyRepository,
+    private val errorFactory: ErrorFactory
+) : UseCase.UseCaseWithParam<String?, List<GiphyDomainModel>> {
 
     override suspend fun executeAsync(query: String?): Resource<List<GiphyDomainModel>> {
         return try {
-             val trendingList = giphyRepository.getTrendingGiphy(query)
-             if (trendingList.isEmpty())  return Resource.empty(errorFactory.createEmptyErrorMessage())
-            //need to map the model to the UI Model
-             val mappedList = trendingList
-             Resource.success(mappedList)
+            val trendingList = giphyRepository.getTrendingGiphy(query)
+            if (trendingList.isEmpty()) return Resource.empty(errorFactory.createEmptyErrorMessage())
+            Resource.success(trendingList)
         } catch (ex: Exception) {
             Log.d("UseCaseError", "executeAsync:  $ex")
-             Resource.error(errorFactory.createApiErrorMessage(ex))
+            Resource.error(errorFactory.createApiErrorMessage(ex))
         }
     }
-
-
-
 }
